# docs/spec/04_M1_OpenAPI.yaml
openapi: 3.1.0
info:
  title: 같이놀개 M1 API
  version: 1.3.1-m1
  description: |
    최신 제품 정책과 M1 통합 ERD를 기준으로 작성한 정적 REST 계약 정본이다.
    런타임 /v3/api-docs는 현재 구현 관찰용이며 계약 정본이나 필수 CI Gate가 아니다.
    M1은 폴링 방식이며 WebSocket, Push, 읽음 표시, 메시지 수정·삭제,
    사용자 셋로그 업로드, 회원탈퇴, 욕설 자동 차단 및 AI 대화 검열을 제공하지 않는다.
    최상위 bearerAuth가 적용되는 모든 operation은 인증 실패 시
    401 ErrorEnvelope를 반환한다.
    별도 HTTP Status는 이 정적 계약을 먼저 변경하고 ErrorCode·구현·테스트를 같은
    PR에서 동기화한다. 소유권 실패는 기본 403이며 존재 은닉이 필요한 endpoint만
    명시적으로 404를 사용한다.
servers:
  - url: /
security:
  - bearerAuth: []
...
paths:
...
  /setlogs:
    get:
      tags: [Setlog]
      summary: M1 공통 시드 셋로그 3개 조회
      operationId: listSeedSetlogs
      description: L1 이상 모든 사용자에게 동일한 시드 영상 3개를 반환한다.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 시드 셋로그
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlogListEnvelope'
  /setlogs/{setlogId}/reactions/{type}:
    parameters:
      - name: setlogId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      - name: type
        in: path
        required: true
        schema:
          $ref: '#/components/schemas/ReactionType'
    put:
      tags: [Setlog]
      summary: 셋로그 반응 추가
      operationId: addSetlogReaction
      description: 동일 반응 추가는 멱등 처리한다. CUTE와 LIKE는 동시에 존재할 수 있다.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 반응 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlogReactionEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags: [Setlog]
      summary: 셋로그 반응 취소
      operationId: removeSetlogReaction
      description: |
        같은 버튼을 다시 누르면 클라이언트가 이 API를 호출한다.
        반응이 이미 없어도 200과 reacted=false를 반환하는 멱등 API다.
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '200':
          description: 반응 상태
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SetlogReactionEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
  /setlogs/{setlogId}/greetings:
    post:
      tags: [Greeting]
      summary: 고정 인사 전송 및 DIRECT 방 생성
      operationId: sendGreeting
      description: |
        요청 본문은 없다. 서버가 "안녕하세요! 같이 놀아요." TEXT를 저장한다.
        상대가 답변하기 전에는 추가 메시지를 보낼 수 없다.
        하루 10명 제한은 Asia/Seoul 날짜 경계로 계산한다.
      parameters:
        - name: setlogId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '401':
          $ref: '#/components/responses/Unauthorized'
        '201':
          description: 인사 및 방 생성 완료
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GreetingEnvelope'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/TooManyRequests'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 보호 API의 인증 실패는 401 ErrorEnvelope로 반환한다.
...
  responses:
...
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Forbidden:
      description: 권한 또는 정책 위반
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    NotFound:
      description: 리소스 없음
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    Conflict:
      description: 상태 또는 중복 충돌
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
    TooManyRequests:
      description: 일일 제한 초과
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorEnvelope'
  schemas:
    ErrorEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success:
          type: boolean
          const: false
        message:
          type: string
        data:
          type: 'null'
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
            message:
              type: string
...
    PetSearchItem:
      type: object
      required: [petId, publicTag, nickname, verified, relationship]
      properties:
        petId:
          type: integer
          format: int64
        publicTag:
          type: string
          maxLength: 30
          pattern: '^.{1,25}#[A-Z0-9]{4}$'
          description: Pet 공개 태그
        nickname:
          type: string
          minLength: 1
        profileUrl:
          type: [string, 'null']
          format: uri
          description: M1은 null이며 클라이언트가 기본 이미지를 표시한다.
        verified:
          type: boolean
        relationship:
          type: [string, 'null']
          enum: [NONE, REQUEST_SENT, REQUEST_RECEIVED, FRIEND, null]
          description: Active Pet이 없는 L1은 null
...
    ReactionType:
      type: string
      enum: [CUTE, LIKE]
    Setlog:
      type: object
      required: [setlogId, authorPet, mediaUrl, mediaUrlExpiresAt, caption, cuteCount, likeCount, myReactions, createdAt]
      properties:
        setlogId:
          type: integer
          format: int64
        authorPet:
          $ref: '#/components/schemas/PetSearchItem'
        mediaUrl:
          type: string
          format: uri
          description: 비공개 S3 객체를 재생하기 위한 Presigned GET URL
        mediaUrlExpiresAt:
          type: string
          format: date-time
          description: 만료 후 GET /setlogs를 다시 호출해 URL을 갱신한다.
        caption:
          type: [string, 'null']
        cuteCount:
          type: integer
          minimum: 0
        likeCount:
          type: integer
          minimum: 0
        myReactions:
          type: array
          uniqueItems: true
          items:
            $ref: '#/components/schemas/ReactionType'
        canInteract:
          type: boolean
          description: L2, 자기 소유 아님, 차단 관계 아님일 때 true
        createdAt:
          type: string
          format: date-time
    SetlogReaction:
      type: object
      required: [setlogId, type, reacted, cuteCount, likeCount]
      properties:
        setlogId:
          type: integer
          format: int64
        type:
          $ref: '#/components/schemas/ReactionType'
        reacted:
          type: boolean
        cuteCount:
          type: integer
        likeCount:
          type: integer
    Greeting:
      type: object
      required: [greetingId, roomId, status, fixedMessage, expiresAt, createdAt]
      properties:
        greetingId:
          type: integer
          format: int64
        roomId:
          type: integer
          format: int64
        status:
          type: string
          enum: [SENT, RESPONDED, EXPIRED]
        fixedMessage:
          type: string
          const: 안녕하세요! 같이 놀아요.
        expiresAt:
          type: string
          format: date-time
        createdAt:
          type: string
          format: date-time
...
    SetlogListEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data:
          type: array
          minItems: 0
          maxItems: 3
          items: { $ref: '#/components/schemas/Setlog' }
        error: { type: 'null' }
    SetlogReactionEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/SetlogReaction' }
        error: { type: 'null' }
    GreetingEnvelope:
      type: object
      required: [success, message, data, error]
      properties:
        success: { type: boolean, const: true }
        message: { type: string }
        data: { $ref: '#/components/schemas/Greeting' }
        error: { type: 'null' }
...